test(sdk-coin-arbeth): update test cases using abstract-eth TICKET: COIN-1322
