+++ modules/abstract-eth/src/test-utils/unit/transactionBuilder/index.ts
@@ -1,0 +1,3 @@
+export * from './addressInitialization';
+export * from './send';
+export * from './walletInitialization';
