+++ modules/sdk-coin-arbeth/test/unit/transferBuilder.ts
@@ -8,12 +8,14 @@ describe('Arbeth send multi sig builder', function () {
   const xprv =
     'xprv9s21ZrQH143K3D8TXfvAJgHVfTEeQNW5Ys9wZtnUZkqPzFzSjbEJrWC1vZ4GnXCvR7rQL2UFX3RSuYeU9MrERm1XBvACow7c36vnz5iYyj2';
   const key = new KeyPair({ prv: xprv }).getKeys().prv as string;
+  const coinName = testData.COIN;
+  const networkTokenIdentifier = testData.NETWORK_TOKEN_IDENTIFIER;
   const amount = '100000000000000000'; // equivalent to 0.1 ether
-  const coin = coins.get('tarbeth') as unknown as EthLikeNetwork;
+  const coin = coins.get(coinName) as unknown as EthLikeNetwork;
   describe('should build', () => {
     it('native coin transfer should succeed', async () => {
       const builder = new TransferBuilder()
-        .coin('tarbeth')
+        .coin(coinName)
         .expirationTime(1590078260)
         .amount(amount)
         .to(toAddress)
@@ -26,7 +28,7 @@ describe('Arbeth send multi sig builder', function () {
 
     it('native coin transfer with sequenceId zero should succeed', async () => {
       const builder = new TransferBuilder()
-        .coin('tarbeth')
+        .coin(coinName)
         .expirationTime(1590078260)
         .amount(amount)
         .to(toAddress)
@@ -39,7 +41,7 @@ describe('Arbeth send multi sig builder', function () {
 
     it('native coin transfer with amount 0 should succeed', async () => {
       const builder = new TransferBuilder()
-        .coin('tarbeth')
+        .coin(coinName)
         .expirationTime(1590078260)
         .amount('0')
         .to(toAddress)
@@ -52,7 +54,7 @@ describe('Arbeth send multi sig builder', function () {
 
     it('ERC20 token transfer should succeed', async () => {
       const builder = new TransferBuilder()
-        .coin('tarbeth:link')
+        .coin(networkTokenIdentifier)
         .expirationTime(1590078260)
         .amount(amount)
         .to(toAddress)
@@ -64,7 +66,7 @@ describe('Arbeth send multi sig builder', function () {
 
     it('erc20 transfer should succeed', async () => {
       const builder = new TransferBuilder()
-        .coin('tarbeth:link')
+        .coin(networkTokenIdentifier)
         .expirationTime(1590078260)
         .amount(amount)
         .to(toAddress)
@@ -76,7 +78,7 @@ describe('Arbeth send multi sig builder', function () {
 
     it('erc20 transfer with amount 0 should succeed', async () => {
       const builder = new TransferBuilder()
-        .coin('tarbeth:link')
+        .coin(networkTokenIdentifier)
         .expirationTime(1590078260)
         .amount('0')
         .to(toAddress)
@@ -99,7 +101,7 @@ describe('Arbeth send multi sig builder', function () {
 
     it('should build from a non signed serialized data', () => {
       const builder = new TransferBuilder(testData.SEND_FUNDS_NO_KEY_DATA);
-      builder.coin('tarbeth').key(key);
+      builder.coin(coinName).key(key);
       const result = builder.signAndBuild(`${coin.chainId}`);
       should.equal(result, testData.SEND_FUNDS_DATA);
     });
